Rename component file for rendering questions
Renamed questionInputBox file to questionBox due to previous name being too specific after refactor included handling of both click and typing questions.
